requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-environ==0.7.0
djangorestframework==3.12.4
ijson==3.1.4
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
celery==5.1.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
psycopg2-binary==2.9.1
django-redis==5.0.0
redis==3.5.3
boto3==1.18.18
pyAesCrypt==6.0.0
django-storages==1.11.1
django-webshell==0.3.0
PyJWT==1.7.1
django-extensions==3.1.3
cryptography==3.4.8
requests==2.26.0
tenacity==8.0.1
django-request-logging==0.7.3
google-api-core==2.4.0
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
